pyproject.toml
[project]
name = "pretalx-plausible"
dynamic = ["version"]
description = "pretalx plugin for pretalx plausible plugin"
readme = "README.rst"
license = {text = "Apache Software License"}
keywords = ["pretalx"]
authors = [
  {name = "Jessica Ringelstein", email = "jessy@jringelstein.de"},
]
maintainers = [
  {name = "Jessica Ringelstein", email = "jessy@jringelstein.de"},
]

requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
dev = [
  "djhtml",
  "pytest",
  "pytest-cov",
  "pytest-django",
  "ruff",
]

[project.urls]
homepage = "https://github.com/JRingelstein/pretalx-plausible"
repository = "https://github.com/JRingelstein/pretalx-plausible"

[project.entry-points."pretalx.plugin"]
pretalx_plausible = "pretalx_plausible:PretalxPluginMeta"

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "pretalx_plausible.__version__"}

[tool.setuptools.packages.find]
include = ["pretalx*"]

[tool.check-manifest]
ignore = [".*", "tests/*", "justfile"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "pretalx.common.settings.test_settings"
addopts = "--cov=pretalx_plausible --cov-report=term-missing:skip-covered --cov-config=pyproject.toml"

[tool.coverage.run]
branch = true
relative_files = true
source = ["pretalx_plausible"]
omit = ["*/migrations/*", "*/urls.py", "*/tests/*"]

[tool.coverage.report]
exclude_also = [
  "def __repr__",
  "def __str__",
  "if settings.DEBUG",
  "NOQA",
  "noqa",
  "NotImplementedError",
]

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN",     # We do not do annotations
  "ARG",     # Unused method arguments are often ok
  "C901",    # We do do complexity
  "COM812",  # Trailing commas are formatter territory
  "D",       # We do not enforce docstyles
  "DJ001",   # We live in a state of sin
  "E501",    # Long lines are formatter territory
  "EM",      # Formatting strings is fine, actually
  "FBT",     # Booleans are fine, actually
  "FIX",     # TODO comments are fine, actually
  "PLR",     # Refactoring: Too many branches, arguments, statements, etc
  "RET502",  # Just returning is fine, actually
  "RET503",  # Just not returning is fine, actually
  "RUF012",  # Mutable default in class attribute is just how django rolls
  "RUF100",  # Unused noqa; problematic when run in isolation
  "TRY003",  # Error messages are fine, actually
]
allowed-confusables = ["–", "'", "×"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = [
  "DTZ",     # TODO: do not use naive datetimes in tests
  "INP001",  # Tests don't need __init__.py
  "RUF",     # Ignore ambiguous characters in tests, may be security tests
  "S",       # Tests do not need cryptographic randomness and other security features
  "S101",    # Ignore `assert` rules in tests
  "SLF001",  # TODO: remove SLF001, fix private access
  "TRY",     # Exception handling is special in tests
]

[tool.ruff.lint.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.lint.flake8-self]
extend-ignore-names = ["_base_manager", "_default_manager",  "_meta"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false
section-order = ["future", "standard-library", "third-party", "pretalx", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
pretalx = ["pretalx"]

[tool.ruff.lint.pep8-naming]
extend-ignore-names = [
  "allDay",
  "api_urls",
  "EventSettings",
  "extendMarkdown",
  "get_allDay",
  "orga_urls",
  "SendMailException",
  "setUpTestData",
  "urls",
]
